# [P4_Symbolic] Optimize the Z3 expressions by factorizing the guard in case of conditionals and tables.