--- a/p4_symbolic/sai/sai_coverage_goals.h
+++ b/p4_symbolic/sai/sai_coverage_goals.h
@@ -10,66 +10,70 @@ inline CoverageGoals SaiDefaultCoverageGoals() {
   return gutil::ParseProtoOrDie<CoverageGoals>(
       R"pb(
         coverage_goals {
-          # TODO: Restore header coverage when performance is
-          # improved.
-          # header_coverage {
-          #  headers { patterns: [ "*" ] }
-          #  header_exclusions {
-          #    patterns: [
-          #      # Optimization: do not need to target ethernet individually.
-          #      All
-          #      # valid packets for SAI-P4 will have the ethernet header.
-          #      "ethernet"
-          #    ]
-          #    patterns: [
-          #      # PacketIO is currently handled differently in dataplane tests.
-          #      "packet_in_header",
-          #      "packet_out_header"
-          #    ]
-          #    patterns: [
-          #      # The following are not satisfiable anyway (because the headers
-          #      # will never be valid in ingress).
-          #      "erspan_ipv4",
-          #      "erspan_gre",
-          #      "erspan_ethernet",
-          #      "tunnel_encap_ipv6",
-          #      "tunnel_encap_gre"
-          #    ]
-          #  }
-          #  headers_to_prevent_unless_explicitly_covered {
-          #    patterns: [ "vlan" ]
-          #  }
-          #  include_wildcard_header: true
-          # }
-          packet_fate_coverage { fates: [ DROP, NOT_DROP ] }
-          entry_coverage {
-            tables { patterns: [ "*" ] }
-            table_exclusions {
-              # TODO: Excluding tables that are covered by covering
-              # other tables and for which packet synthesis is expensive.
-              patterns: [
-                "ingress.routing_resolution.neighbor_table",
-                "ingress.routing_resolution.nexthop_table",
-                "ingress.routing_resolution.router_interface_table",
-                "ingress.routing_resolution.wcmp_group_table",
-                "ingress.routing_resolution.tunnel_table",
-                # TODO: Remove the following when PINS releases no
-                # longer include the tables.
-                "ingress.routing.neighbor_table",
-                "ingress.routing.nexthop_table",
-                "ingress.routing.router_interface_table",
-                "ingress.routing.wcmp_group_table",
-                "ingress.routing.tunnel_table"
-              ]
+          cartesian_product_coverage {
+            # TODO: Restore header coverage when performance is
+            # improved.
+            # header_coverage {
+            #  headers { patterns: [ "*" ] }
+            #  header_exclusions {
+            #    patterns: [
+            #      # Optimization: do not need to target ethernet individually.
+            #      All
+            #      # valid packets for SAI-P4 will have the ethernet header.
+            #      "ethernet"
+            #    ]
+            #    patterns: [
+            #      # PacketIO is currently handled differently in dataplane
+            #      tests.
+            #      "packet_in_header",
+            #      "packet_out_header"
+            #    ]
+            #    patterns: [
+            #      # The following are not satisfiable anyway (because the
+            #      # headers will never be valid in ingress).
+            #      "erspan_ipv4",
+            #      "erspan_gre",
+            #      "erspan_ethernet",
+            #      "tunnel_encap_ipv6",
+            #      "tunnel_encap_gre"
+            #    ]
+            #  }
+            #  headers_to_prevent_unless_explicitly_covered {
+            #    patterns: [ "vlan" ]
+            #  }
+            #  include_wildcard_header: true
+            # }
+            packet_fate_coverage { fates: [ DROP, NOT_DROP ] }
+            entry_coverage {
+              tables { patterns: [ "*" ] }
+              table_exclusions {
+                # TODO: Excluding tables that are covered by
+                # covering other tables and for which packet synthesis is
+                # expensive.
+                patterns: [
+                  "ingress.routing_resolution.neighbor_table",
+                  "ingress.routing_resolution.nexthop_table",
+                  "ingress.routing_resolution.router_interface_table",
+                  "ingress.routing_resolution.wcmp_group_table",
+                  "ingress.routing_resolution.tunnel_table",
+                  # TODO: Remove the following when GPINS releases
+                  # no longer include the tables.
+                  "ingress.routing.neighbor_table",
+                  "ingress.routing.nexthop_table",
+                  "ingress.routing.router_interface_table",
+                  "ingress.routing.wcmp_group_table",
+                  "ingress.routing.tunnel_table"
+                ]
+              }
+              cover_default_actions: true
+              # TODO: Currently must ignore cover for default
+              # action of empty tables due to the limitation of
+              # packet_util::SynthesizePackets's return type
+              # (TestPacketsByTargetEntry).
+              # TODO: Also covering these tables is
+              # expensive. Remove when these limitations are addressed.
+              exclude_empty_tables: true
             }
-            cover_default_actions: true
-            # TODO: Currently must ignore cover for default
-            # action of empty tables due to the limitation of
-            # packet_util::SynthesizePackets's return type
-            # (TestPacketsByTargetEntry).
-            # TODO: Also covering these tables is
-            # expensive. Remove when these limitations are addressed.
-            exclude_empty_tables: true
           }
         }
       )pb");