--- a/p4_symbolic/symbolic/guarded_map.cc
+++ b/p4_symbolic/symbolic/guarded_map.cc
@@ -69,6 +69,8 @@ absl::Status SymbolicGuardedMap::Set(absl::string_view key, z3::expr value,
   // This will return an absl error if the sorts are incompatible, and will pad
   // shorter bit vectors.
   ASSIGN_OR_RETURN(old_value, operators::Ite(guard, value, old_value));
+  // Simplifying the guard every time the map is updated.
+  old_value = old_value.simplify();
   return absl::OkStatus();
 }